+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.4.0 (2019-06-05)
+* Add `OTR::ActiveRecord::QueryCache` middleware.
+
+### 1.3.0
+* Active Record 6.0 support
+
 ### 1.2.7 (2019-01-11)
 * A less hacky way of fixing the bug fixed in 1.2.6. `ENV["RACK_ENV"]`/`ENV["RAILS_ENV"]` will no longer be forced into `development` when blank.